move simulateWriteContractAndNotifyError to commom util
