+++ packages/nextjs/hooks/scaffold-eth/useScaffoldWriteContract.ts
@@ -2,11 +2,12 @@ import { useEffect, useState } from "react";
 import { MutateOptions } from "@tanstack/react-query";
 import { Abi, ExtractAbiFunctionNames } from "abitype";
 import { Config, UseWriteContractParameters, useAccount, useConfig, useWriteContract } from "wagmi";
-import { WriteContractErrorType, WriteContractReturnType, simulateContract } from "wagmi/actions";
+import { WriteContractErrorType, WriteContractReturnType } from "wagmi/actions";
 import { WriteContractVariables } from "wagmi/query";
 import { useSelectedNetwork } from "~~/hooks/scaffold-eth";
 import { useDeployedContractInfo, useTransactor } from "~~/hooks/scaffold-eth";
-import { AllowedChainIds, getParsedError, notification } from "~~/utils/scaffold-eth";
+import { AllowedChainIds, notification } from "~~/utils/scaffold-eth";
+import { simulateContractWriteAndNotifyError } from "~~/utils/scaffold-eth/common";
 import {
   ContractAbi,
   ContractName,
@@ -83,16 +84,6 @@ export function useScaffoldWriteContract<TContractName extends ContractName>(
     chainId: selectedNetwork.id as AllowedChainIds,
   });
 
-  const simulateAndNotifyError = async (params: WriteContractVariables<Abi, string, any[], Config, number>) => {
-    try {
-      await simulateContract(wagmiConfig, params);
-    } catch (error) {
-      const parsedError = getParsedError(error);
-      notification.error(parsedError);
-      throw error;
-    }
-  };
-
   const sendContractWriteAsyncTx = async <
     TFunctionName extends ExtractAbiFunctionNames<ContractAbi<TContractName>, "nonpayable" | "payable">,
   >(
@@ -125,7 +116,7 @@ export function useScaffoldWriteContract<TContractName extends ContractName>(
       } as WriteContractVariables<Abi, string, any[], Config, number>;
 
       if (!finalConfig?.disableSimulate) {
-        await simulateAndNotifyError(writeContractObject);
+        await simulateContractWriteAndNotifyError({ wagmiConfig, writeContractParams: writeContractObject });
       }
 
       const makeWriteWithParams = () =>
